requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic
pymongo
gunicorn
motor
pandas
pydantic[email]